Altered the linefeed function to take three arguments
Also removed some debuggin output
